move interpreter tests to a new file
