+++ src/test/interpreter_tests.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace Test {
+struct TestSet;
+}
+
+void interpreter_tests(Test::TestSet& tests);
